requirements.txt
openai
langchain
langchain-core
langchain-openai
langchain-community
langchain-anthropic
langchain-google-genai
langchain-text-splitters
langchain_experimental
langchain-huggingface
langchain-chroma
google-generativeai
transformers
huggingface-hub
python-dotenv
numpy
scikit-learn
streamlit
pydantic
pydantic[email]
grandalf
pypdf
bs4
chromadb
tiktoken
faiss-cpu 
wikipedia
youtube-transcript-api